+++ src/backend/mailbox_manager/tests/fixtures/dimail.py
@@ -16,7 +16,11 @@ CHECK_DOMAIN_BROKEN = {
     "smtp_domain": None,
     "context_name": "example.fr",
     "transport": None,
-    "domain_exist": {"ok": True, "internal": False, "errors": []},
+    "domain_exist": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
     "mx": {
         "ok": False,
         "internal": False,
@@ -74,8 +78,16 @@ CHECK_DOMAIN_BROKEN = {
             {"code": "no_dkim", "detail": "Il faut un DKIM record, avec la bonne clef"}
         ],
     },
-    "postfix": {"ok": True, "internal": True, "errors": []},
-    "ox": {"ok": True, "internal": True, "errors": []},
+    "postfix": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+    "ox": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
     "cert": {
         "ok": False,
         "internal": True,
@@ -84,6 +96,141 @@ CHECK_DOMAIN_BROKEN = {
         ],
     },
 }
+
+
+CHECK_DOMAIN_BROKEN_INTERNAL = {
+    "name": "example.fr",
+    "state": "broken",
+    "valid": False,
+    "delivery": "virtual",
+    "features": ["webmail", "mailbox"],
+    "webmail_domain": None,
+    "imap_domain": None,
+    "smtp_domain": None,
+    "context_name": "example.fr",
+    "transport": None,
+    "domain_exist": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "mx": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "cname_imap": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "cname_smtp": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "cname_webmail": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "spf": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "dkim": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "postfix": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+    "ox": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+    "cert": {
+        "ok": False,
+        "internal": True,
+        "errors": [
+            {"code": "no_cert", "detail": "Pas de certificat pour ce domaine (ls)"}
+        ],
+    },
+}
+
+CHECK_DOMAIN_BROKEN_EXTERNAL = {
+    "name": "example.fr",
+    "state": "broken",
+    "valid": False,
+    "delivery": "virtual",
+    "features": ["webmail", "mailbox"],
+    "webmail_domain": None,
+    "imap_domain": None,
+    "smtp_domain": None,
+    "context_name": "example.fr",
+    "transport": None,
+    "domain_exist": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "mx": {
+        "ok": False,
+        "internal": False,
+        "errors": [
+            {
+                "code": "wrong_mx",
+                "detail": "Je veux que le MX du domaine soit mx.ox.numerique.gouv.fr., or je trouve example-fr.mail.protection.outlook.com.",
+            }
+        ],
+    },
+    "cname_imap": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "cname_smtp": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "cname_webmail": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "spf": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "dkim": {
+        "ok": True,
+        "internal": False,
+        "errors": [],
+    },
+    "postfix": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+    "ox": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+    "cert": {
+        "ok": True,
+        "internal": True,
+        "errors": [],
+    },
+}
+
 
 CHECK_DOMAIN_OK = {
     "name": "example.fr",